% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gamm_rates.R
\name{calc_gamm_rates}
\alias{calc_gamm_rates}
\title{Calculate station- or site-level rates of change from SET or MH data using
generalized additive mixed models (GAMM).}
\usage{
calc_gamm_rates(
  data,
  level = "station",
  kval = 10,
  override_site_corrections = FALSE
)
}
\arguments{
\item{data}{data frame. A data frame of raw SET or MH data. See details below
for requirements.}

\item{level}{string (optional). Level at which to calculate rates of surface
  elevation change or vertical accretion. One of:
  \itemize{
    \item "station": (default) station-level rates of surface elevation change.
    \item "site": site-level rates of surface elevation change.
}}

\item{kval}{The basis dimension for the smooth term for time; defaults to 10.
See \code{\link[mgcv]{choose.k}}.}

\item{override_site_corrections}{logical (optional). If TRUE, don't use the
site groupings specified in utils.R since these are specific to the USGS
project.}
}
\value{
For SET data, returns a data frame of station- or site-level rates
  of surface elevation change. For MH data, returns a data frame of station-
  or site-level rates of surface accretion. Note that units for rates of
  change are mm/yr.
}
\description{
This function takes a data frame of raw SET or MH data and calculates a rate
of surface elevation change (for SET data) or vertical accretion (for MH
data) at either the station- or site-level using a generalized additive mixed
models (GAMM) model as implemented in the package \pkg{mgcv}.
}
\details{
For station-level SET data, random effect smooths are calculated for
  each pin nested within SET direction. For site-level SET data, random
  effect smooths are calculated for each pin nested within each SET direction
  nested within each SET station. For station-level MH data, random effect
  smooths are calculated for each duplicate marker horizon (i.e., plot A, B,
  C, etc.). For site-level MH data, random effect smooths are calculated for
  each duplicate marker horizon nested within each SET station. For MH data,
  duplicate core readings ('core_measurement_number') are first averaged to
  the marker horizon-level. GAMM models use thin-plate regression splines
  s(bs = "tp") for the smooth term for time and random effects s(bs = "re")
  for the various random effect terms - see \code{\link[mgcv]{smooth.terms}}.
  Note that rates of change, standard errors, and confidence intervals are
  calculated as the mean of 200-equally spaced first derivatives of the
  function representing the smooth term for time via the 'derivatives'
  function from the gratia package - see \code{\link[gratia]{derivatives}}.
}
\note{
Stations with fewer than 3 measurement dates are excluded from the
  calculation of rates.
}
\section{Data Requirements}{


  This function takes a data frame of either raw SET or MH data (`data`).

  SET data must have 1 row per pin reading and the following columns, named
  exactly: event_date_UTC, network_code, park_code, site_name, station_code,
  SET_direction, pin_position, SET_offset_mm, pin_length_mm, and
  pin_height_mm. Note that SET_offset_mm and pin_length_mm can be empty (aka
  blank) but the columns must be included in the data frame. See
  `data(example_sets)`.

  MH data must have 1 row per core measurement and the following columns,
  named exactly: event_date_UTC, network_code, park_code, site_name,
  marker_horizon_name, core_measurement_number, core_measurement_depth_mm,
  and established_date. See `data(example_mh)`.
}

\examples{
# Defaults to station-level rates
calc_gamm_rates(example_sets)

# Site-level cumulative change
calc_gamm_rates(example_mh, level = "site")

# Can also be used with pipes to include/exclude specific stations, dates, etc.
## exclude station M11-3 from calculations of gamm rates
example_sets \%>\%
    filter(station_code != "M11-3") \%>\%
    calc_gamm_rates(., level = "site")

## Exclude data from after 2016 from calculations of gamm rates
example_sets \%>\%
    filter(event_date_UTC < as.Date("2016-01-01")) \%>\%
    calc_gamm_rates(., level = "station")

## Define custom groups for calculating gamm rates
example_sets \%>\%
    group_by(set_type) \%>\%
    calc_gamm_rates(., level = "site")

}
\references{
Feher, L.C., Osland, M.J., Johnson, D.J., Grace, J.B.,
  Guntenspergen, G.R., Stewart, D.R., Coronado-Molina, C., and Sklar, F.H.,
  2024. Nonlinear patterns of surface elevation change in coastal wetlands:
  the value of generalized additive models for quantifying rates of change.
  Estuaries and Coasts, 47:1893-1902.
  \url{https://doi.org/10.1007/s12237-023-01268-w}
}
\seealso{
\code{\link{calc_lmm_rates}},
  \code{\link{calc_gam_rates}}, \code{\link[mgcv]{gam}},
  \code{\link[gratia]{derivatives}}
}
