% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lmm_rates.R
\name{calc_lmm_rates}
\alias{calc_lmm_rates}
\title{Calculate station- or site-level rates of change from SET or MH data using
generalized linear mixed models}
\usage{
calc_lmm_rates(
  data,
  level = "station",
  override_site_corrections = FALSE,
  drop_SET_direction = FALSE,
  drop_station_code = FALSE
)
}
\arguments{
\item{data}{data frame. A data frame of raw SET or MH data. See details below
for requirements.}

\item{level}{string (optional). Level at which to calculate rates of surface
  elevation change or vertical accretion. One of:
  \itemize{
    \item "station": (default) station-level rates of surface elevation change.
    \item "site": site-level rates of surface elevation change.
}}

\item{override_site_corrections}{logical (optional). If TRUE, don't use the
site groupings specified in utils.R since these are specific to the USGS
project.}

\item{drop_SET_direction}{boolean. Optional; for use with level = "station"
if the model(s) fail to converge.}

\item{drop_station_code}{boolean. Optional; for use with level = "site" if
the model(s) fail to converge.}
}
\value{
For SET data, returns a data frame of station- or site-level rates
  of surface elevation change. For MH data, returns a data frame of station-
  or site-level rates of surface accretion. Note that units for rates of
  change are mm/yr.
}
\description{
This function takes a data frame of raw SET or MH data and calculates a rate
of surface elevation change (for SET data) or vertical accretion (for MH
data) at either the station- or site-level using a generalized linear mixed
model as implemented in the package \pkg{nlme}.
}
\details{
For station-level SET data, random intercepts are calculated for
  each pin nested within SET direction. For site-level SET data, random
  intercepts are calculated for each pin nested within each SET direction
  nested within each SET station. For station-level MH data, random
  intercepts are calculated for each duplicate marker horizon (i.e., plot A,
  B, C, etc.). For site-level MH data, random intercepts are calculated for
  each duplicate marker horizon nested within each SET station. For MH data,
  duplicate core readings ('core_measurement_number') are first averaged to
  the marker horizon-level.
}
\note{
Stations with fewer than 3 measurement dates are excluded from the
  calculation of rates.
}
\section{Data Requirements}{


  This function takes a data frame of either raw SET or MH data (`data`).

  SET data must have 1 row per pin reading and the following columns, named
  exactly: event_date_UTC, network_code, park_code, site_name, station_code,
  SET_direction, pin_position, SET_offset_mm, pin_length_mm, and
  pin_height_mm. Note that SET_offset_mm and pin_length_mm can be empty (aka
  blank) but the columns must be included in the data frame. See
  `data(example_sets)`.

  MH data must have 1 row per core measurement and the following columns,
  named exactly: event_date_UTC, network_code, park_code, site_name,
  marker_horizon_name, core_measurement_number, core_measurement_depth_mm,
  and established_date. See `data(example_mh)`.
}

\examples{
# Defaults to station-level rates
calc_lmm_rates(example_sets)

# Site-level rates
calc_lmm_rates(example_mh, level = "site")

# Can also be used with pipes to include/exclude specific stations, dates, etc.
## exclude station M11-3 from calculations of rates
example_sets \%>\%
    filter(station_code != "M11-3") \%>\%
    calc_lmm_rates(., level = "site")

## Exclude data from after 2016 from calculations of rates
example_sets \%>\%
    filter(event_date_UTC < as.Date("2016-01-01")) \%>\%
    calc_lmm_rates(., level = "station")

## Define custom groups for calculating rates
example_sets \%>\%
    group_by(set_type) \%>\%
    calc_lmm_rates(., level = "site")

}
\seealso{
\code{\link{calc_linear_rates}}, \code{\link[nlme]{lme}}
}
