% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mh}
\alias{example_mh}
\title{Example MH data}
\format{
A data frame with 664 rows and 9 variables:
\describe{
  \item{event_date_UTC}{Date UTC; measurement date, yyyy-mm-dd format}
  \item{network_code}{chr;, 4-letter I&M network code}
  \item{park_code}{chr; 4-letter NPS park code}
  \item{site_name}{chr; site name}
  \item{station_code}{chr; station code}
  \item{marker_horizon_name}{chr; marker horizon replicate name}
  \item{core_measurement_number}{dbl; number of the measurement taken from a single core}
  \item{core_measurement_depth_mm}{dbl; measured depth to the marker horizon in mm}
  \item{established_date}{Date UTC; date that the marker horizon plot was established, yyyy-mm-dd format}
}
}
\usage{
example_mh
}
\description{
A sample of marker horizon data from M11 at ASIS
}
\keyword{datasets}
