% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_slr_rate.R
\name{get_long_slr_rate}
\alias{get_long_slr_rate}
\title{Get the current park-specific "long-term" rate of sea-level rise published by
NOAA tides and currents}
\usage{
get_long_slr_rate(park, nauset = FALSE)
}
\arguments{
\item{park}{A string (required). A capitalized four-letter park code. One of:
\itemize{
  \item "ACAD": Acadia National Park - gauge 8413320 Bar Harbor, ME.
  \item "BOHA": Boston Harbor Islands National Recreation Area - gauge 8443970 Boston, MA.
  \item "CACO": Cape Cod National Seashore - 8443970 (Boston, MA) or 8449130 for Nauset (Nantucket Island, MA).
  \item "FIIS": Fire Island National Seashore - gauge 8531680 Sandy Hook, NJ.
  \item "GATE": Gateway National Recreation Area - gauge 8531680 Sandy Hook, NJ.
  \item "NACE": National Capital Area East - gauge 8594900 Washington, DC.
  \item "GWMP": George Washington Memorial Parkway - gauge 8594900 Washington, DC.
  \item "ASIS": Assateague Island National Seashore - gauge 8570283 (Ocean City Pier, MD) for long-term/recent SLR, gauge 8570280 (Ocean City Inlet, MD) for future SLR .
  \item "COLO": Colonial National Historical Park - gauge 8638610 Sewells Point, VA.
  \item "CAHA": Cape Hatteras National Seashore - gauge 8652587 (Oregon Inlet Marina, NC) for long-term/recent SLR, gauge 8654400 (Cape Hatteras, NC) for future SLR.
  \item "CALO": Cape Lookout National Seashore - gauge 8656483 Beaufort, NC.
  \item "FOPU": Fort Pulaski National Monument - gauge 8670870 Fort Pulaski, GA.
  \item "FOFR": Fort Frederica National Monument - gauge 8720030 Fernandina Beach, GA.
  \item "CUIS": Cumberland Island National Seashore - gauge 8720030 Fernandina Beach, GA.
  \item "TIMU": Timucuan Ecological and Historic Preserve - gauge 8720218 Mayport, FL (Bar Pilots Dock).
  \item "FOMA": Fort Matanzas National Monument - gauge 8720218 Mayport, FL (Bar Pilots Dock).
  \item "CANA": Canaveral National Seashore - gauge 8721604 Trident Pier, FL.
  \item "BISC": Biscayne National Park - gauge 8723214 Virginia Key, FL.
  \item "VIIS": Virgin Islands National Park - gauge 9751639 Charlotte Amalie, St. Thomas, USVI.
  \item "SARI": Salt River Bay National Historic Park and Ecological Preserve - gauge 9751401 Lime Tree Bay, St. Croix, USVI.
}}

\item{nauset}{Logical (optional). Used to select the Nauset gauge for CACO;
defaults to FALSE.}
}
\value{
A data frame with the current park-specific "long-term" rate of
  sea-level rise published by NOAA.
}
\description{
This function downloads the current "long-term" rate of sea-level rise
published on the NOAA tides and currents website for the nearest tide gauge
(\url{https://tidesandcurrents.noaa.gov/sltrends/}).
}
\details{
Rates of sea-level rise reported by NOAA are (generally) updated on
  a regular basis. The rates of sea-level rise (here referred to as the
  long-term rate) are based on the entire period of record for each NOAA
  gauge, however, the gauges vary significantly in age and therefore this
  "long-term" rate may be based on a relatively short time period for some
  gauges. For example, the Ocean City MD gauge was installed in 1975 and thus
  the period of record for the gauge is 1975 - present (50 years in 2025), in
  contrast to the Boston gauge which was installed in 1921 and thus has a
  period of record of 1921 - present (104 years in 2025).
}
\examples{
get_long_slr_rate(park = "ASIS")

}
\seealso{
\code{\link{get_recent_slr_rate}}, \code{\link{get_sea_level_data}}
}
