% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mh_data.R
\name{load_mh_data}
\alias{load_mh_data}
\title{Download/Load raw marker horizon data from the NPS SET data package, NPS SET
database, or a saved file}
\usage{
load_mh_data(
  park = NULL,
  network_code = NULL,
  file_path = NULL,
  db_server = NULL,
  db_port = NULL
)
}
\arguments{
\item{park}{string (optional); A 4 character park code (e.g., "ASIS", "GATE")
that can be used to filter results from the NPS SET data package or NPS SET
database to a specific park. If not specified, will return data for all
parks.}

\item{network_code}{string (optional); A 4 character network code (e.g.,
"NCBN", "NCRN") that can be used to filter results from the NPS SET data
package or NPS SET database to a specific I&M network. If not specified,
will return data for all networks.}

\item{file_path}{string (optional); Instead of downloading the most recent
SET data package or connecting to the NPS SET database, you can supply a
file path to load data from a saved .csv, .xlsx, or .xls file. Can supply
either a full file path or use `here::here` for a relative file path. See
details below for requirements.}

\item{db_server}{string (optional); A connection string for accessing the NPS
I&M SET database back-end.}

\item{db_port}{number (optional); A port value for accessing the NPS I&M SET
database back-end.}
}
\value{
A data frame of raw MH data.
}
\description{
This function can be used to download and/or load raw marker horizon data.
Defaults to downloading data from the most recent SET data package if neither
`file_path =` or `db_server =` are specified. Uses functions from the
\pkg{NPSutils} package to download data packages.
}
\details{
Note that you must be on VPN and have access to the back-end of NPS
  I&M SET database for the database option to work. Contact Laura Feher
  (lfeher at NPS.gov) for the required connection strings.
}
\note{
If loading in your own data from csv, see
  \code{\link{calc_change_cumu}} for data requirements.
}
\examples{
\dontrun{
# Download and load raw MH data for ASIS from the most recent data package
df <- load_mh_data(park = "ASIS")

# Download and load raw MH data for NCBN parks from the online database
df <- load_mh_data(network_code = "NCBN", db_server = "example.server")

# Load raw MH data from a saved csv file
df <- load_mh_data(file_path = "C:/Documents/Data/my_MH_data.csv")

# Use `here::here` to load raw MH data from a relative file path
df <- load_mh_data(file_path = here::here("Data", "my_MH_data.csv"))
}

}
\seealso{
\code{\link{load_set_data}}
}
