% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_comps.R
\name{plot_rate_comps}
\alias{plot_rate_comps}
\title{Plot comparison of station- or site-level surface elevation change or
vertical accretion rates}
\usage{
plot_rate_comps(data = NULL, rates = NULL, level = "station")
}
\arguments{
\item{data}{A data frame (optional). A data frame of raw SET or MH data. See
details below for requirements.}

\item{rates}{A data frame (optional). A data frame of station- or site-level
rates of surface elevation change. See details below for requirements.}

\item{level}{string (optional). Level at which to calculate rates of surface
  elevation change or vertical accretion. One of:
  \itemize{
    \item "station": (default) station-level rates of surface elevation change.
    \item "site": site-level rates of surface elevation change.
}}
}
\value{
a ggplot object: depending on the type of data supplied, x-axis
  values are rates of either (a) cumulative surface elevation change or (b)
  vertical accretion. If `level = "station"` (default), the y-axis is station
  codes. If `level = "site"`, the y-axis is site names.
}
\description{
This function creates a 'TIE fighter' style plot that can be used to visually
compare station- or site-level linear rates of surface elevation change or
vertical accretion.
}
\details{
Accepts either a data frame of either raw SET or MH data (`data` argument) or
a data frame of pre-calculated station- or site-level rates of surface
elevation change or vertical accretion (`rates` argument).
}
\note{
Cumulative change is calculated via the function
  \code{\link{calc_change_cumu}} and linear rates of change are calculated
  via the function \code{\link{calc_linear_rates}}. See function
  documentation for details.
}
\section{Data Requirements for pre-calculated rates of change}{


  This function takes a data frame of either raw SET data (`data`) or a
  user-created data frame of station- or site-level rates of surface
  elevation change (`rates`).

  If supplying a user created data frame to  `rates` and `level = station`,
  must be a data frame of station-level rates of surface elevation change or
  vertical accretion with 1 row per station and the following columns, named
  exactly: station_code, rate, and rate_se. The "rate" and "rate_se" columns
  should represent the station-level rates and their standard errors
  (respectively).

  If supplying a user created data frame to `rates` and `level = site`, must
  be a data frame of site-level rates of surface elevation change or vertical
  accretion with 1 row per site and the following columns, named exactly:
  site_name, rate, and rate_se. The "rate" and "rate_se" columns should
  represent the site-level rates and their standard errors (respectively).
}

\section{Data Requirements}{


  This function takes a data frame of either raw SET or MH data (`data`).

  SET data must have 1 row per pin reading and the following columns, named
  exactly: event_date_UTC, network_code, park_code, site_name, station_code,
  SET_direction, pin_position, SET_offset_mm, pin_length_mm, and
  pin_height_mm. Note that SET_offset_mm and pin_length_mm can be empty (aka
  blank) but the columns must be included in the data frame. See
  `data(example_sets)`.

  MH data must have 1 row per core measurement and the following columns,
  named exactly: event_date_UTC, network_code, park_code, site_name,
  marker_horizon_name, core_measurement_number, core_measurement_depth_mm,
  and established_date. See `data(example_mh)`.
}

\section{Details}{


  For SET data, pin-level cumulative change is first calculated as the
  difference between each pin reading and the reading from the earliest date
  that was not NA. Note that for data from NCBN, NCRN, and NETN, pin heights
  are first converted to standardized pin heights using the formula 1000 +
  (SET_offset_mm - (pin_length_mm - pin_height_mm)) to account for the 6"
  extensions that were added to some benchmarks within these networks' parks
  in 2023/24. Then, cumulative pin changes are averaged to the arm-level on
  each date, and then the cumulative arm-level changes are averaged up to the
  station-level on each date. If level = "site" is used, station-level
  cumulative change will be averaged up to the site-level.

  For MH data, duplicate core readings ('core_measurement_number') are first
  averaged to the marker horizon-level. Then, average marker horizon depths
  are averaged up to the station-level on each date, excluding NAs. If level
  = "site" is used, station-level cumulative change will be averaged up to
  the site-level.
}

\examples{

# Default plot gives station-level rates
plot_rate_comps(data = example_sets)

# Example with a user-supplied data frame of station-level rates
example_rates <- data.frame("station" = c("station_1", "station_2", "station_3"),
                            "rate" = c(3.2, 4.0, 5.4),
                            "se_rate" = c(1, 0.5, 0.25))

plot_rate_comps(rates = example_rates)

# Site-level rates
plot_rate_comps(data = example_sets, level = "site")

}
\seealso{
\code{\link{calc_change_cumu}}, \code{\link{calc_linear_rates}},
  \code{\link{plot_cumu}}
}
