% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_slr_rate.R
\name{get_recent_slr_rate}
\alias{get_recent_slr_rate}
\title{Calculate a park-specific recent rate of sea-level rise using sea-level data
from the nearest NOAA tide gauge.}
\usage{
get_recent_slr_rate(park, nauset = FALSE, start_year = 2001, end_year = 2019)
}
\arguments{
\item{park}{A string (required). A capitalized four-letter park code. One of:
* `"ACAD"`: Acadia National Park - gauge 8413320 Bar Harbor, ME.
* `"BOHA"`: Boston Harbor Islands National Recreation Area - gauge 8443970 Boston, MA.
* `"CACO"`: Cape Cod National Seashore - 8443970 (Boston, MA) or 8449130 for Nauset (Nantucket Island, MA).
* `"FIIS"`: Fire Island National Seashore - gauge 8531680 Sandy Hook, NJ.
* `"GATE"`: Gateway National Recreation Area - gauge 8531680 Sandy Hook, NJ.
* `"NACE"`: National Capital Area East - gauge 8594900 Washington, DC.
* `"GWMP"`: George Washington Memorial Parkway - gauge 8594900 Washington, DC.
* `"ASIS"`: Assateague Island National Seashore - gauge 8570283 (Ocean City Pier, MD) for long-term/recent SLR, gauge 8570280 (Ocean City Inlet, MD) for future SLR .
* `"COLO"`: Colonial National Historical Park - gauge 8638610 Sewells Point, VA.
* `"CAHA"`: Cape Hatteras National Seashore - gauge 8652587 (Oregon Inlet Marina, NC) for long-term/recent SLR, gauge 8654400 (Cape Hatteras, NC) for future SLR.
* `"CALO"`: Cape Lookout National Seashore - gauge 8656483 Beaufort, NC.
* `"FOPU"`: Fort Pulaski National Monument - gauge 8670870 Fort Pulaski, GA.
* `"FOFR"`: Fort Frederica National Monument - gauge 8720030 Fernandina Beach, GA.
* `"CUIS"`: Cumberland Island National Seashore - gauge 8720030 Fernandina Beach, GA.
* `"TIMU"`: Timucuan Ecological and Historic Preserve - gauge 8720218 Mayport, FL (Bar Pilots Dock).
* `"FOMA"`: Fort Matanzas National Monument - gauge 8720218 Mayport, FL (Bar Pilots Dock).
* `"CANA"`: Canaveral National Seashore - gauge 8721604 Trident Pier, FL.
* `"BISC"`: Biscayne National Park - gauge 8723214 Virginia Key, FL.
* `"VIIS"`: Virgin Islands National Park - gauge 9751639 Charlotte Amalie, St. Thomas, USVI.
* `"SARI"`: Salt River Bay National Historic Park and Ecological Preserve - gauge 9751401 Lime Tree Bay, St. Croix, USVI.}

\item{nauset}{Logical (optional). Used to select the Nauset gauge for CACO;
defaults to FALSE.}

\item{start_year}{A four digit year (optional). The starting year that you
want to get sea-level data for. If start_year is not supplied (default),
returns data for the entire period of record for that station.}

\item{end_year}{An four digit year (optional). The ending year that you want
to get sea-level data for. If end_year is not supplied (default), returns
data for the entire period of record for that station.}
}
\value{
A data frame with a park-specific recent rate of sea-level rise
  calculated using the sea-level data from the nearest NOAA tide gauge for
  the specified time period.
}
\description{
This function downloads sea-level data for the nearest NOAA tide gauge from
the tides and currents website (https://tidesandcurrents.noaa.gov/sltrends/)
and then calculates a rate of sea-level rise for the specified time period.
Defaults to the 19 year time period for the most recent tidal epoch:
2001-2019.
}
\details{
The recent rate of sea-level rise is calculated via linear
  regression using monthly mean sea-levels with the regular seasonal
  fluctuations removed. Sea-level data is downloaded from the NOAA tides and
  currents website https://tidesandcurrents.noaa.gov/. Note that monthly mean
  sea level values are relative to the most recent mean sea-level datum
  established by CO-OPS
  (https://tidesandcurrents.noaa.gov/datum_options.html).The value in the
  'estimate' column for the 'yr' term represents the recent rate of sea-level
  rise, and the value in 'std.error' column represents the standard error of
  the recent sea-level rise rate.
}
\examples{
get_recent_slr_rate(park = "ASIS")

}
\references{
Moon, J.A., Feher, L.C., Lane, T.C., Vervaeke, W.C., Osland,
  M.J., Head, D.M., Chivoiu, B.C., Stewart, D.R., Johnson, D.J., Grace, J.B.,
  and Metzger, K.L., 2022. Surface elevation change dynamics in coastal
  marshes along the Northwestern Gulf of Mexico: Anticipating effects of
  rising sea-level and intensifying hurricanes. Wetlands
  42(5):49.(https://doi.org/10.1007/s13157-022-01565-3)

Chivoiu, B., Osland, M.J., Collini, R., Martin, S., Tirpak, J.,
  and Wilson, B., 2020. Local sea level rise information sheets for Texas,
  Louisiana, Mississippi, Alabama and Florida: Northern Gulf of Mexico
  Sentinel Site Cooperative, U.S. Geological Survey and U.S. Fish and
  Wildlife Service: Mississippi-Alabama Sea Grant Consortium,
  (https://placeslr.org/our-products/federally-managed-lands-two-pagers/)
}
\seealso{
[get_sea_level_data()]
}
